refactor(test): Move specs to root unit folder
