+++ test/unit/assistant/dates.spec.js
@@ -2,10 +2,10 @@ const path = require("path");
 const fsExtra = require("fs-extra");
 const sinon = require("sinon");
 
-const { setDate } = require("../../../../src/assistant/dates");
-const { readExifDates } = require("../../../../src/exif/fileMethods");
-const { setLevel, _logger } = require("../../../../src/support/tracer");
-const { formatForLogsFromExif } = require("../../../../src/support/dates");
+const { setDate } = require("../../../src/assistant/dates");
+const { readExifDates } = require("../../../src/exif/fileMethods");
+const { setLevel, _logger } = require("../../../src/support/tracer");
+const { formatForLogsFromExif } = require("../../../src/support/dates");
 
 const {
   assetPath,
@@ -14,7 +14,7 @@ const {
   resetTempPath,
   copyAssetToTempPath,
   TEMP_PATH,
-} = require("../../../support/assets");
+} = require("../../support/assets");
 
 setLevel("silent");
