+++ test/unit/assistant/setDatesFixturesIso.spec.js
@@ -1,6 +1,6 @@
 const fsExtra = require("fs-extra");
 
-const { setDatesInFolder } = require("../../../../src/assistant/runner");
+const { setDatesInFolder } = require("../../../src/assistant/runner");
 
 const {
   resetTempPath,
@@ -9,8 +9,8 @@ const {
   tempOutputPath,
   copyFixturesToTempPath,
   tempFixturesFolder,
-} = require("../../../support/assets");
-const { readExifDates } = require("../../../../src/exif/fileMethods");
+} = require("../../support/assets");
+const { readExifDates } = require("../../../src/exif/fileMethods");
 
 const FIXTURE = "iso";
