+++ test/unit/assistant/setDatesFixturesOneFormat.spec.js
@@ -1,4 +1,4 @@
-const { setDatesInFolder } = require("../../../../src/assistant/runner");
+const { setDatesInFolder } = require("../../../src/assistant/runner");
 
 const {
   resetTempPath,
@@ -6,17 +6,17 @@ const {
   tempOutputPath,
   copyFixturesToTempPath,
   tempFixturesFolder,
-} = require("../../../support/assets");
-const { readExifDates } = require("../../../../src/exif/fileMethods");
+} = require("../../support/assets");
+const { readExifDates } = require("../../../src/exif/fileMethods");
 
-const FIXTURE = "iso-regex";
+const FIXTURE = "one-format";
 
 describe(`setDates executed in ${FIXTURE} fixtures`, () => {
   beforeAll(async () => {
     await resetTempPath();
   });
 
-  describe("when modify option is true, setDigited is false and dateRegex is provided", () => {
+  describe("when modify option is true, setDigited is false and format is provided", () => {
     beforeAll(async () => {
       await resetTempPath();
       await copyFixturesToTempPath(FIXTURE);
@@ -25,53 +25,46 @@ describe(`setDates executed in ${FIXTURE} fixtures`, () => {
         outputFolder: TEMP_OUTPUT_FOLDER,
         fromDigitized: false,
         modify: true,
-        dateRegex: /^date\-(\S*)/,
+        dateFormat: "dd_MM_yyyy",
       });
     });
 
     it("should have set date from file name to image with date in root folder", async () => {
-      const { DateTimeOriginal } = await readExifDates(
-        tempOutputPath("date-2021-09-25T122030.jpg")
-      );
-      expect(DateTimeOriginal).toEqual(`2021:09:25 12:20:30`);
+      const { DateTimeOriginal } = await readExifDates(tempOutputPath("25_09_2021.jpg"));
+      expect(DateTimeOriginal).toEqual(`2021:09:25 00:00:00`);
     });
 
     it("should have set date from file name to image with date in subfolder", async () => {
       const { DateTimeOriginal } = await readExifDates(
-        tempOutputPath("subfolder", "date-2021-09.jpeg")
+        tempOutputPath("subfolder", "01_09_2021.jpeg")
       );
       expect(DateTimeOriginal).toEqual(`2021:09:01 00:00:00`);
     });
 
     it("should have set date from file name to image with date in sub subfolder", async () => {
       const { DateTimeOriginal } = await readExifDates(
-        tempOutputPath("subfolder", "date-2013-10-23T124015", "date-2020-10-09T172815.jpeg")
+        tempOutputPath("subfolder", "23_10_2013", "09_10_2020.jpeg")
       );
-      expect(DateTimeOriginal).toEqual(`2020:10:09 17:28:15`);
+      expect(DateTimeOriginal).toEqual(`2020:10:09 00:00:00`);
     });
 
     it("should have set date from folder name to image with not date in sub subfolder", async () => {
       const { DateTimeOriginal } = await readExifDates(
-        tempOutputPath("subfolder", "date-2013-10-23T124015", "gorilla.JPG")
+        tempOutputPath("subfolder", "23_10_2013", "gorilla.JPG")
       );
-      expect(DateTimeOriginal).toEqual(`2013:10:23 12:40:15`);
+      expect(DateTimeOriginal).toEqual(`2013:10:23 00:00:00`);
     });
 
     it("should have set date from file name to image with date in sub sub subfolder", async () => {
       const { DateTimeOriginal } = await readExifDates(
-        tempOutputPath(
-          "subfolder",
-          "date-2013-10-23T124015",
-          "date-1979-05",
-          "date-2045-10-12.jpeg"
-        )
+        tempOutputPath("subfolder", "23_10_2013", "01_05_1979", "12_10_2045.jpeg")
       );
       expect(DateTimeOriginal).toEqual(`2045:10:12 00:00:00`);
     });
 
     it("should have set date from folder name to image without date in sub sub subfolder", async () => {
       const { DateTimeOriginal } = await readExifDates(
-        tempOutputPath("subfolder", "date-2013-10-23T124015", "date-1979-05", "gorilla.JPG")
+        tempOutputPath("subfolder", "23_10_2013", "01_05_1979", "gorilla.JPG")
       );
       expect(DateTimeOriginal).toEqual(`1979:05:01 00:00:00`);
     });
