+++ test/unit/index.spec.js
@@ -1,4 +1,4 @@
-const index = require("../../../src/index");
+const index = require("../../src/index");
 
 describe("Index file", () => {
   describe("run method", () => {
